Implement multi line logs.
To do so we parse the log message splitting on \n. The first line is logged normally, subsequent lines are logged on a new line with "|    ", i.e | followed by four spaces. We use strtok_rm and not strktok, because of thread safety. For portability to windows we #define strtok_r strtok_s because they are the same thing and MSFT just named it differently. This commit also adds a function which replaces certain special characters in the logging with doubly escaped versions, e.g we log \r as \\r so we can clearly see it.