ssh/scp to deploy CI build, use ping to poll progress
Rewrote the deployment script to not use the endpoint on the server, because that end point requires the GHA get artifacts endpoint and the API does not list artifacts for a run until that run has completed. The logic is replicated in a new script deploy_artifact_ssh.sh, which uses ssh and scp to stop the server, copy files and restart the server. I also added a ping with timeout to the deploy to VM step. This will ping for 300 seconds, or until it receives a packet back, which ever comes first.