Begin integrating sentry into protocol client
Also fixes the server build to link to the correct libcrypto: system for linux server, our bundled version for windows.
Modify logging macros to send sentry breadcrumbs for warnings and events for errors.
Set up sentry in init logger and destroy in log destroy.
Add definition for sentry environment, one of dev, staging, or production. This is intended to be set from a cmake flag e.g -DFRACTAL_ENVIRONMENT="dev", this is used to group sentry events.