Run get_latest_sentry.py directly from the workflows.
We are doing this because we want to use an authenticated API request and passing the github token around from github actions to cmake to python is more trouble than it is worth. Devs will need to run get_latest_sentry.py before compiling the first time they clone the repo.