WIP Sentry set user email on client and server, tag connection id in client and server.
This commit adds a new FractalServerMessageType: MESSAGE_USER_EMAIL this sends the users email to the server. The users email is supplied to the client through a new optional command line argument, -u or --user. This is sent to sentry on both sides. On the server, the host email is sent to sentry as a user, additional spectators are sent as breadcrumbs (logs). We also send the connection id generated by the server, which was already sent to the client, as a sentry tag. This will allow us to match server and client events in the sentry dashboard.