WIP sentry set environment
This commit adds the -e --environment command line argument to set the environment for the client. On the server we use the dev environment until we update the webserver parameters and then if it is in dev mode we set the environment to dev and otherwise to the branch returned by the server request.