The beginning of self-managed cluster scaling (#528)
Added function to start self-managing cluster scaling. Currently just checks for scale-down when a container is deleted and if criteria is met, performs the scale down. Criteria is simple: 1) if the number of instances * AWS_TASKS_PER_INSTANCE (config db) > active_tasks + pending_tasks, and 2) if there are instances with 0 tasks, then perform a scale-down.