:adhesive_bandage: fix building in github workflow
the first issue was caused because we linked the input
library to the client app. the second was simply because
our github actions runners were trying to build the server
on 18.04. the third issue was that bumping to 20.04 also
bumped the version of cppcheck, which was too aggressive on
lodepng, which is an external library, and with some of the
header files in the external dependencies for sentry-native.
the fourth issue was that clang-tidy on ubuntu 20.04 is by
default version 10, which is far more aggressive than the 6.0
which we have been testing against. the fifth issue was that
clang-tidy-6.0 needs to be symlinked to clang-tidy in order
for our scripts to be able to call it.