:heavy_division_sign: factor out the grid driver installation
We do this by moving it to another stage in the multi-stage build. We further minify the installation options and configure all relevant files to be installed under the prefix `/dist`. Therefore, the main-stage build can simply `COPY --from=nvidia-grid-driver /dist /` to install all necessary components of the Nvidia grid driver.

In the future, this will enable us to get even more granular about which components of the driver we install, and which or whether we may mount certain components from the host system.

In terms of image size, this takes the chrome image from `2.91 gb` to `2.57 gb`. Not the most massive change, I know, but still fairly significant for such a simple fix.