Actually block IMDS
According to https://aws.amazon.com/premiumsupport/knowledge-center/ecs-container-ec2-metadata/#Resolution, we actually need to do this on the [`DOCKER-USER` chain](https://serverfault.com/a/933803), not the `FORWARD` one to "avoid [our] rules being clobbered by docker".

I've tested that this blocks the issue from FCTL-20 / #57 and that this doesn't completely brick networking from inside containers. As long as this doesn't break anything, I would like to get this merged so that it's really simple to test, since there seem to be differences between our deployment instances and manually-spun instances with respect to networking security.