`yarn lint:{check,fix}` sets appropriate exit code
In the old client app, this wasn't actually possible because `yarn` wouldn't properly set this exit code (we were directly running the CLI and combining it with prettier). Even with a bash script, it wasn't clean to get this exit code working, so all of our workflows instead would detect errors based on the `std{out,err}` of the linting command.

But no more! We can now just grab the exit code as the return value of the (async) CLI entrypoint, and exit with that exit code. This will make our linting workflows appropriately fail on error.