Safely exit application before container delete request
We do this by using `systemctl stop fractal-display`. This sends a `SIGTERM` to the application in question, followed by waiting for up to 90 seconds and then a `SIGKILL`. It's possible that the 90 seconds will be an issue, but fortunately this `systemd` behavior is highly configurable. The `SIGTERM`, when it is handled correctly, will cleanly exit the application and allow it to finish normal shutdown procedures. The `SIGKILL` meanwhile will continue to force-shutdown the application, thus continuing current behavior as a fallback on error.