Preliminary implementation of nacking for frames which are missing in their entirety.
Current implementation:
- Added a new timer to the VideoData struct for keeping track of how
often we're nacking for missing frames
- If any of the frames in the ring buffer between the next frame to
render and the maximum packet ID received are stale, this means that
we're missing all the packets for those frames, so we nack for the
index 0 packet.
- We don't nack if the last rendered id is really far away from the
max ID - this probably indicates that the client is starting up, and
would result in too many nacks to the server.