Append timestamp to all info sent to logzio via Info func
The code uses the golang package logzio/logzio-go which will collect the messages in a queue and upload them to logzio in batches based on a default drain duration. (https://github.com/logzio/logzio-go/blob/master/logziosender.go#L325). One option of solving this problem was to update the drain duration to very little ie 5 sec -> 1 sec but that may increase the requests me make by x5 and still will not be precise. This option will show the time in the message although it will not match the timestamp show on the web UI.