update ami active status before draining instances
we want to swap the order to updating new ami's protected and active
status + current ami active status before draining. There is a chance
draining fails and causes the swap to fail. We want to make sure the
statuses are updated first to prevent the issue above. We will add a
new piece to the clean up script to drain instances on old ami's to
catch instances that failed to drain.