Avoid the worst case of a race to destroy clipboard synchronizer
The detached thread races with the destruction of the mutex and associated
variables in the caller, which was segfaulting for me on Linux.  This
change ensures that we wait for the currently-active clipboard thread to
finish before destroying, though if very unlucky there could still be
previously-detached clipboard threads hanging around to break things.