Never detach threads in clipboard synchronizer
Detaching threads lead to a race on cleanup when resources like the
mutex are destroyed but there is no way to tell whether the threads
have all finished.  This instead keeps a list of outstanding threads,
and when possible it joins those which have marked an atomic variable
to say they are definitely finished (so that joining doesn't block).
On final cleanup, it then blocks to join all outstanding threads
before destroying anything else.