Atomic clipboard (don't detach threads) (#4898)
* Build with C11

Needed for standard atomic support.  This should have no effect on all
existing code, which was previously built with C99.

* Add header for atomic operations

Use C11 atomics if possible, which should be available everywhere
except MSVC.  On MSVC, provide a simple implementation of the default
sequentially-consistent operations using the Interlocked* functions.
Use C++11 atomics matching C11 behaviour when building with C++.

* Never detach threads in clipboard synchronizer

Detaching threads lead to a race on cleanup when resources like the
mutex are destroyed but there is no way to tell whether the threads
have all finished.  This instead keeps a list of outstanding threads,
and when possible it joins those which have marked an atomic variable
to say they are definitely finished (so that joining doesn't block).
On final cleanup, it then blocks to join all outstanding threads
before destroying anything else.

* Add test for atomic operations

Runs multiple threads doing many simultaneous adds and subtracts on an
atomic variable which should exactly cancel out.

* Add const to thread name argument to thread creation

Avoids warnings when passing a string literal.