Fix Video pixelization issue
The fix is done via two major changes

Change 1:
Limiting the max Inter QP to 30 and Intra QP to 45 to avoid the video pixelation issue at low bitrates.
Sometimes congestion control chooses a really low bitrate at which the visual quality is unbearable.
To avoid that we set a lower bound on video quality, even if it means exceeding the bitrate chosen by congestion control algorithm.
Higher the QP, Lower the bitrate and Lower the visual quality. Allowed QP range is 0-51

Change 2:
Change 1 means that there is a risk of bitrate overflow and client not getting any frame at all.
To avoid this risk we drop the input frames even before encoding whenever the bitrate overflows.
Average Bitrate is calculated over the VBV Buffer size interval.

Minor Fix 1:
There is also another minor bug in i-frame insertion that got exposed due to this change.
Same frames got encoded in IPP structure at a rate greater than 60 fps due to this bug.
This bug is also fixed in this commit.