Fix some problems found with sanitisers (#5027)
* lodepng: Fix memory leak

lodepng_decode32() returns a heap pointer which needs to be freed.

* logger: Fix memory leak

* memory: Make it possible to destroy a block allocator

So that we can fix memory leaks elsewhere.  Also avoid the awkward
indirection through InternalBlockAllocator - we can just define the
structure in the private file.

* ringbuffer: Fix memory leaks

Ensure that all buffers have been returned to the packet buffer allocator
and then free it.

* logging: Fix memory leak

The crash handler won't work at this point anyway because it depends on
the already-destroyed logger mutexes.

* protocol_test: Fix memory leaks

* keyboard_mapping: Fix signed overflow

Found by ubsan.

* protocol_test: Fix signed overflow

Found by ubsan.