Avoid frame drops in capture
When 60 fps video such as this[https://www.youtube.com/watch?v=vX2vsvdq8nw] is played 6-7 fps is dropped at the capture logic itself.
Even during normal browsing, 2-3 fps is dropped at the capture logic itself.
This is because we are trying to the time the call to capture screen at 1/60 seconds.
This timing invariably goes wrong it causes us to call capture screen late causing frame drops.

In this commit, we just call capture screen during every iteration of the loop and it anyways returns 0, if no new frames are available.
Thus we capture and encode a frame as soon as it is available.