Reduce the CPU usage of WhistServer process
multithreaded_send_video is busy looping hogging up the CPU core even when it is not doing any useful work.
A sleep of 0.1ms is added to the loop, whenever there are no frames to encode or send.
The CPU usage of WhistServer process comes down drastically from 80-90% to 35-40% with this change.
This change doesn't have any impact on FPS.