Avoid unexpected EINTR returns on receive
Add new functions recv_no_intr() and recvfrom_no_intr() which include
the do-while-not-EINTR loop, and use them to replace bare recv() and
recvfrom() calls which do not want to see EINTR returns.

This includes support for socket receive timeouts - we check for whether
one is present after an EINTR return, and if so account for it by
calling poll() for the time remaining on the socket (possibly
repeatedly, if more signals arrive).