Fix protocol testing framework to open the configured URL
The protocol testing framework was not opening any URL and was just in Chrome new-tab window for the entire test duration.
This PR ensures that the new tab URL is passed on to the client via -x (--new-tab-url) option.
Also fixed a bug in client that was causing stack overflow error when using --new-tab-url option.