Move video packet send to a separate thread
Video capture + encode consumes an average of 10ms or so.
For a video running at 60 FPS, that leaves only 6.66ms per frame for sending the video packets.
This means we can utilize only 6.66/16.66 = 40% of the available bandwidth for sending video packets.
This becomes a major bottleneck during low bandwidth situations, where we are not able to utilize the full bandwidth efficiently and end up dropping frames.

Moving network send to a different thread is a commonly used design to solve this problem.
When video encode of frame 'n' is happening, video packets of frame 'n-1' is sent over the network in parallel.