Make the video output at a constant 60 fps
We need to maintain a constant FPS for the video encoder to produce the bitrate requested by the client. Maintaining the requested video encoder bitrate is very important for any congestion control algorithm to work correctly.
Currently the FPS is around 45-50, for normal browsing.
Another advantage of this FPS increase is reduced latency and better user experience.
But the disadvantage of this change is the increased CPU usage of WhistClient from 29% to 35%, and hence a slightly higher power consumption. But this is a small price to pay for a stable congestion control, reduced latency and a much better user experience.