Define MSVC option to disable C runtime warnings globally
This is the warning which tells you to use the incompatible Windows-only
versions of C11 Annex K functions instead of normal standard library
ones, which we never want.  Set the _CRT_SECURE_NO_WARNINGS define
globally so that we don't need to set it individually in files which
might be used on Windows.