General code cleanup 2 (#5374)
* Disable MSVC zero-sized struct element warning globally

This is used in many places in the code; ensuring we have included a
header which disables it is not helpful.

* Define MSVC option to disable C runtime warnings globally

This is the warning which tells you to use the incompatible Windows-only
versions of C11 Annex K functions instead of normal standard library
ones, which we never want.  Set the _CRT_SECURE_NO_WARNINGS define
globally so that we don't need to set it individually in files which
might be used on Windows.

* Move internal-only parts of clipboard handling to a new file

Ensure that all the `unsafe_` internal functions are using the same
prototypes.

* Rewrite getopt_long() header without LGPL

This was previously using the LGPL header from the GNU C library, which
imposed licensing issues we were not compliant with.  Rewrite to only
include the things we need, using the manpages as reference.

* Move internal-only parts of input device handline to a new file

* Fix remaining prototypes in whist common code

* Fix prototypes in server code

* Fix prototypes in client code

* Enable prototype-checking options with gcc and clang