Fix thread-unsafe logging of timestamps (#5448)
* Remove explicit time strings from log lines

The timestamp is already included in the log context information.

* Fix thread-unsafe logging of timestamps

Change current_time_str() to write into a buffer given by the caller
rather than using a single static buffer.  Update logging to use that,
which necessitates moving the writing of the log line context
information into the logging function.  Also reduce the amount of
copying and avoid a heap allocation if possible.

* Threaded logging test

Output log lines in a tight loop on four threads simultaneously.