File capture device (#5426)
* Move protocol test assets into a single versioned archive

This makes it possible to change files without breaking builds in other
branches.

* Add new capture device which reads from a file

This reads the file with libavformat and decodes the contents with
libavcodec.  It also scales with libswscale to the size desired by the
user if necessary.  Includes tests for running it with an MP4 file and
with a single JPEG (which is returned repeatedly).