Use DPI as well for bitrate limit calculation
Last week a PR was sent to adjust bitrate limits as per the screen resolution.
But when the same limits are used for a screen with lesser DPI the visual quality is very bad.
So this PR uses DPI as well for bitrate limit calculations for a better visual quality.