Better thresholds for stream reset logic
- Acceptable staleness threshold calculated as follows
Acceptable staleness = Time to transmit that frame + 1 RTT(to recover from any nacks) + Network Jitter
Here Network Jitter is assumed to be 0.3 times of RTT, as used in nacking logic currently.
And this value is capped between the range of 100ms to 200ms to ensure that this doesn't go too far from the current value of 100ms.

- MAX_UNSYNCED_FRAMES increased from 4 to 6.
The current value of 4 was decided when the MIN_FPS was at 45.
Now that the MIN_FPS is increased to 60, MAX_UNSYNCED_FRAMES also needs to be increased to maintain similar duration as threshold.