UDP ping refactor and 4-ping disconnect fix
Current UDP code was ignoring the pongs received later than 500ms and marking this as failure.
And 4 such consecutive failures are marked as server disconnected.
But this is not correct as pongs received later doesn't mean connection is lost. Connection is still alive.

This commit refactors the code to handle the pongs received after the next ping was sent(basically later than 500ms).
And UDP connection is marked as failure, only if no pong is received for 5000ms(to bring it on par with TCP timeout)