Don't throttle audio packets
Throttling audio packets is unnecessary for the following reasons.
- Audio encoder runs on a reserved bandwidth for itself, unlike video encoder doesn't exceed its reserved bitrate significantly.
- Audio frame sizes are very small. At 128 Kbps, the average Audio frame size is only 160 bytes, much lesser than an average UDP packet size.
- Audio frames at produced at a near constant interval of 10ms

Throttling audio packets is counter-productive for the following reason.
Throttler uses a burst interval of 5ms. For sending the current audio packet + 1 prev audio packet, we can easily exceed the 10ms audio frame during times of congestion.
Unlike a video encoder, audio encoder runs on a tight frame duration of exactly 10ms.
Any drift in audio fps on the server side due to congestion will easily cause the audio buffer to underflow, followed by an overflow on the client side.
So throttling audio is causing a lot of stutters during congestion