WCC tuning for large queue networks
For large queue length networks, congestion causes huge latency before showing up as packet loss.
The current WCC algo is not detecting this kind of congestion early resulting in higher latencies.
In this change we will detect congestion by comparing the short term average of latency with its long term average.
If the latency's short-term average increases beyond a particular threshold relative to the long-term average then it is now detected as congestion.