Prevent a screen resize(in linux) from half resolution to full resolution immediately after startup
We always start the window in maximized mode. But we are also trying to set the unmaximize resolution to half the "maximize" resolution.
But these parameters are ignored in Mac and unmaximize(double click the title bar) doesn't work until we manually resize the window atleast once.
On linux it is causing the window to start with half resolution and then maximize it 100ms or so later, causing barrage of stream resets during startup.

So this change is a no-op in Mac and performance improvement in Linux. Required for E2E testing as it runs in linux.