Partially fix RingBuffer overflow error
The "Ringbuffer overflowed; stream resets have been failing to recover." error log is generally seen during stream freeze.
But it is also seen when an "empty frame" is lost during screen idle time.
This is because stream resets are not honored immediately during the idle time.
This change will just restore the previous behaviour of sending a recovery frame immediately, even in idle phase.