Fix "The capture session must be recreated" sentry error
As per the NVFBC API doc,
NVFBC_ERR_MUST_RECREATE - This indicates that the application must recreate the capture session.
This error can be returned if a modeset event occurred while capturing frames, and NVFBC_CREATE_- HANDLE_PARAMS::bDisableAutoModesetRecovery was set to NVFBC_TRUE.

7.3.2.2 NVFBC_BOOL _NVFBC_CREATE_CAPTURE_SESSION_- PARAMS::bDisableAutoModesetRecovery
[in] Whether NvFBC should not attempt to recover from modesets.
NvFBC is able to detect when a modeset event occured and can automatically re-create a capture session with the same settings as before, then resume its frame capture session transparently.
This option allows to disable this behavior. NVFBC_ERR_MUST_RECREATE will be returned in that case.
It can be useful in the cases when an application needs to do some work between setting up a capture and grabbing the first frame.

Since this is not applicable for us, let us set bDisableAutoModesetRecovery to FALSE, for transparent handling of this corner-case issue.