Serialise video frames
This splits apart the network and the in-memory representation of video
frames, which allows us to save a lot of network bytes in every frame
sent since we no longer send any fields which are not needed.

It also fixes UBSan errors for unaligned access due to invalid aliasing
of the VideoFrame structure.